refactor: 각 탭의 Content 컴포넌트로 분리
