+++ components/post/tabs/index.tsx
@@ -1,26 +1,10 @@
 import styled from '@emotion/styled';
 import * as Tabs from '@radix-ui/react-tabs';
-import React, { useState } from 'react';
 import { COLORS } from 'styles/colors';
-import RefreshIcon from 'public/icons/circle_arrow.svg';
-import CommentList from './CommentList';
-import ParticipantList from './ParticipantList';
-import { usePostComment } from '../hooks/queries';
-import { useRouter } from 'next/router';
+import CommentTapContent from './comment';
+import ParticipantTapContent from './participant';
 
 export default function PostTabs() {
-  const [commentValue, setCommentValue] = useState('');
-  const { mutate: postComment } = usePostComment();
-  const router = useRouter();
-  const { postId } = router.query;
-
-  const onChangeCommentValue = (e: React.ChangeEvent<HTMLInputElement>) => setCommentValue(e.target.value);
-  const onSubmitComment = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    postComment({ postId: postId as string, comment: commentValue });
-    setCommentValue('');
-  };
-
   return (
     <Tabs.Root defaultValue="comments">
       <Tabs.List>
@@ -28,25 +12,10 @@ export default function PostTabs() {
         <StyledParticipantsTabTrigger value="participants">참여자</StyledParticipantsTabTrigger>
       </Tabs.List>
       <Tabs.Content value="comments">
-        <StyledCommentsTabContainer>
-          <Background>
-            <RefreshButton>
-              <RefreshIcon />
-            </RefreshButton>
-            <CommentList />
-            <CommentForm onSubmit={onSubmitComment}>
-              <CommentInput onChange={onChangeCommentValue} value={commentValue} placeholder="댓글을 입력하세요" />
-              <CommentSubmitButton>완료</CommentSubmitButton>
-            </CommentForm>
-          </Background>
-        </StyledCommentsTabContainer>
+        <CommentTapContent />
       </Tabs.Content>
       <Tabs.Content value="participants">
-        <StyledParticipantsTabContainer>
-          <Background>
-            <ParticipantList />
-          </Background>
-        </StyledParticipantsTabContainer>
+        <ParticipantTapContent />
       </Tabs.Content>
     </Tabs.Root>
   );
@@ -70,7 +39,7 @@ const StyledParticipantsTabTrigger = styled(StyledTabTrigger)`
   background-color: ${COLORS.sub_green};
 `;
 
-const StyledTabContainer = styled.div`
+export const StyledTabContainer = styled.div`
   border-radius: 20px;
   width: 501px;
   height: 726px;
@@ -79,15 +48,7 @@ const StyledTabContainer = styled.div`
   justify-content: center;
 `;
 
-const StyledCommentsTabContainer = styled(StyledTabContainer)`
-  background-color: ${COLORS.main};
-`;
-
-const StyledParticipantsTabContainer = styled(StyledTabContainer)`
-  background-color: ${COLORS.sub_green};
-`;
-
-const Background = styled.div`
+export const TabContentBackground = styled.div`
   background-color: white;
   border-radius: 20px;
   width: 450px;
@@ -95,43 +56,4 @@ const Background = styled.div`
   display: flex;
   flex-direction: column;
   padding: 0 33px;
-`;
-
-const RefreshButton = styled.button`
-  border-radius: 28px;
-  background-color: ${COLORS.main};
-  padding: 8px 15px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  align-self: flex-end;
-  margin-top: 20px;
-`;
-
-const CommentForm = styled.form`
-  position: relative;
-  margin-bottom: 15px;
-`;
-
-const CommentInput = styled.input`
-  border: 2px solid #d9d9d9;
-  border-radius: 11px;
-  padding: 12px 58px 12px 19px;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 21px;
-  color: #838383;
-  width: 100%;
-`;
-
-const CommentSubmitButton = styled.button`
-  font-weight: 700;
-  font-size: 18px;
-  line-height: 21px;
-  color: ${COLORS.main};
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  right: 14px;
-  background-color: transparent;
 `;
